refactor: divide, rename, delete and remove Vehicle related code
Feito múltiplas operações de mover, renomear, apagar e dividir componentes e código a fim de estruturar mais logicamente e nomear mais semanticamente páginas (com o sufixo Page).
